refactor: Move the remainder of utils.search_engine into utils.search and utils.tables
